some improvmetns
